Dockerization
I've only added a simple docker file and gunicorn to the requirements so
that docker can run the Django app. I based the dockerfile off of this
example: https://github.com/yvsssantosh/django-polls-rest/blob/master/Dockerfile

The CSS is broken with this and I'm not sure why. I think something to do
with collectstatic? This is mostly just a base example of what is
required to make the Djanog app run in a container so that we can deploy
to kubernetes.